tools/rag-pipeline/requirements.txt
chromadb-client>=0.5.0
numpy>=1.24.0
httpx>=0.27.0
google-genai>=1.0.0
python-docx>=1.1.0
watchdog>=4.0.0
pytest>=8.0.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
# OCR & Buecher-Pipeline
Pillow>=10.0.0
opencv-python-headless>=4.9.0
pdf2image>=1.17.0
PyMuPDF>=1.24.0
imagehash>=4.3.0
ffmpeg-python>=0.2.0
